requirements.txt
# =============================================================================
# Audio Preprocessing Container - Dependencies
# =============================================================================

# -----------------------------------------------------------------------------
# Azure SDKs
# -----------------------------------------------------------------------------
azure-storage-blob>=12.24.0
azure-storage-queue>=12.12.0
azure-identity>=1.19.0

# -----------------------------------------------------------------------------
# Audio Processing
# -----------------------------------------------------------------------------
# Core audio I/O - uses libsndfile (supports WAV, FLAC, OGG natively)
soundfile>=0.13.0

# Audio analysis and high-quality resampling
librosa>=0.11.0

# Numerical computing
numpy>=2.1.0
scipy>=1.14.0

# -----------------------------------------------------------------------------
# Voice Activity Detection
# -----------------------------------------------------------------------------
# Silero VAD - enterprise-grade, ML-based
silero-vad>=6.0.0

# ONNX Runtime for fast inference (CPU only, smaller package)
onnxruntime>=1.20.0

# -----------------------------------------------------------------------------
# Noise Reduction (optional but recommended for call center audio)
# -----------------------------------------------------------------------------
noisereduce>=3.0.0

# -----------------------------------------------------------------------------
# Web Framework (for health checks, metrics, manual triggers)
# -----------------------------------------------------------------------------
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# -----------------------------------------------------------------------------
# Observability
# -----------------------------------------------------------------------------
# Structured logging
structlog>=24.4.0

# Metrics (Prometheus format)
prometheus-client>=0.21.0

# OpenTelemetry for distributed tracing (optional)
# opentelemetry-api>=1.28.0
# opentelemetry-sdk>=1.28.0
# opentelemetry-exporter-otlp>=1.28.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
# Configuration management
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Retry logic
tenacity>=9.0.0

# Async utilities
aiofiles>=24.1.0

# -----------------------------------------------------------------------------
# Development (excluded in production builds)
# -----------------------------------------------------------------------------
# pytest>=8.3.0
# pytest-asyncio>=0.24.0
# pytest-cov>=6.0.0
# httpx>=0.28.0  # For testing FastAPI
